feat: refactor attendance card
