+++ apps/web/src/app/events/components/AttendanceCard/AttendanceDateInfo.tsx
@@ -4,6 +4,20 @@ import { formatDate } from "date-fns"
 import { nb } from "date-fns/locale"
 import React from "react"
 
+const getFormatString = (isInSameYear: boolean) => (isInSameYear ? "dd. MMMM" : "dd.MM.yyyy")
+
+const formatTime = (date: Date) => formatDate(date, "HH:mm", { locale: nb })
+
+const dateComponent = (label: string, dateStr: string, time: string) => (
+  <div>
+    <Text className="text-base">{label}</Text>
+    <div className="flex flex-row gap-2 text-slate-12 text-base sm:flex-col sm:gap-0 sm:text-sm">
+      <Text>{dateStr}</Text>
+      <Text>kl. {time}</Text>
+    </div>
+  </div>
+)
+
 interface AttendanceDateInfoProps {
   attendance: Attendance
 }
@@ -17,8 +31,6 @@ export const AttendanceDateInfo = ({ attendance }: AttendanceDateInfoProps) => {
   const isAttendanceClosedInSameYear = attendance.registerEnd.getFullYear() === now.getFullYear()
   const isDeregisterDeadlineInSameYear = attendance.deregisterDeadline.getFullYear() === now.getFullYear()
 
-  const getFormatString = (isInSameYear: boolean) => (isInSameYear ? "dd. MMMM" : "dd.MM.yyyy")
-
   const registerStartDate = formatDate(attendance.registerStart, getFormatString(isAttendanceStartInSameYear), {
     locale: nb,
   })
@@ -31,23 +43,11 @@ export const AttendanceDateInfo = ({ attendance }: AttendanceDateInfoProps) => {
     { locale: nb }
   )
 
-  const formatTime = (date: Date) => formatDate(date, "HH:mm", { locale: nb })
-
-  const renderDateBlock = (label: string, dateStr: string, time: string) => (
-    <div>
-      <Text className="text-base">{label}</Text>
-      <div className="flex flex-row gap-2 text-slate-12 text-base sm:flex-col sm:gap-0 sm:text-sm">
-        <Text>{dateStr}</Text>
-        <Text>kl. {time}</Text>
-      </div>
-    </div>
-  )
-
   const dateBlocks = [
     {
       key: "registerStart",
       date: attendance.registerStart,
-      element: renderDateBlock(
+      element: dateComponent(
         isAttendanceStartInPast ? "Åpnet" : "Åpner",
         registerStartDate,
         formatTime(attendance.registerStart)
@@ -56,7 +56,7 @@ export const AttendanceDateInfo = ({ attendance }: AttendanceDateInfoProps) => {
     {
       key: "registerEnd",
       date: attendance.registerEnd,
-      element: renderDateBlock(
+      element: dateComponent(
         isAttendanceClosedInPast ? "Lukket" : "Lukker",
         registerEndDate,
         formatTime(attendance.registerEnd)
@@ -65,7 +65,7 @@ export const AttendanceDateInfo = ({ attendance }: AttendanceDateInfoProps) => {
     {
       key: "deregisterDeadline",
       date: attendance.deregisterDeadline,
-      element: renderDateBlock("Avmeldingsfrist", deregisterDeadlineDate, formatTime(attendance.deregisterDeadline)),
+      element: dateComponent("Avmeldingsfrist", deregisterDeadlineDate, formatTime(attendance.deregisterDeadline)),
     },
   ]
